Micro re-overhaul; applying forces more selectively; controlling target distance; maintaining width/depth. Lazy target selection for performance. Chasing. Build tweaks.

+++ src/Information/Geography/Pathfinding/Types/TilePath.scala
@@ -5,7 +5,7 @@ import Mathematics.Points.Tile
 case class TilePath(
   start     : Tile,
   end       : Tile,
-  distance  : Float,
+  distance  : Double,
   tiles     : Option[IndexedSeq[Tile]]) {
   
   def pathExists: Boolean = tiles.isDefined
